rename urls.ts to navigation.ts
