+++ packages/e2e-tests/tests/page.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { goToFacilityFrontend } from '../utils/urls';
+import { goToFacilityFrontend } from '../utils/navigation';
 
 test('homepage has expected title', async ({ page }) => {
   await goToFacilityFrontend(page);
